requirements.txt
apprise==1.9.1
black==24.10.0
boto3==1.35.85
botocore==1.35.85
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
idna==3.10
jmespath==1.0.1
markdown==3.7
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pycryptodome==3.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyyaml==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.4
setuptools==75.1.0
six==1.17.0
urllib3==2.2.3
wheel==0.44.0
